Inline compute_utils into run_slurm, privatize internal helpers
Folds compute_utils.py (RunSpec, command/script builders, CUDA flags) into
scripts/run_slurm.py — its only caller — and drops the now-unused module.
Also privatizes single-module helpers and removes the unused
avg_eval_metrics_across_ranks.
